--- src/Plugin.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace Cache;
-
-use Cake\Core\BasePlugin;
-
-class Plugin extends BasePlugin {
-
-	/**
-	 * @var bool
-	 */
-	protected bool $middlewareEnabled = false;
-
-	/**
-	 * @var bool
-	 */
-	protected bool $bootstrapEnabled = false;
-
-	/**
-	 * @var bool
-	 */
-	protected bool $routesEnabled = false;
-
-}
